% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_model_identification.R
\name{enhance_step}
\alias{enhance_step}
\title{Title}
\usage{
enhance_step(
  traintest,
  response,
  global,
  xreg_candidates,
  h,
  exclude = exclude_var,
  spare_cores
)
}
\arguments{
\item{spare_cores}{}
}
\description{
Title
}
