% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_model_identification.R
\name{identify_model}
\alias{identify_model}
\title{Title}
\usage{
identify_model(
  data,
  predictors,
  specific = paste0("auftragseingang_", 1:6),
  spare_cores = 1
)
}
\arguments{
\item{specific}{variables that are availaible only for specific prediction steps}

\item{spare_cores}{}
}
\description{
Title
}
