% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_model_identification.R
\name{make_candidates_list}
\alias{make_candidates_list}
\title{create models to test}
\usage{
make_candidates_list(n, existing, new, baseline = NULL)
}
\arguments{
\item{baseline}{}
}
\description{
create models to test
}
