% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_model_identification.R
\name{make_test_fc}
\alias{make_test_fc}
\title{make test forecasts for further evaluation}
\usage{
make_test_fc(traintest, response, xreg, exclude_var)
}
\arguments{
\item{traintest}{a named list with nested lists \strong{train} and \strong{test}}

\item{response}{variable to be predicted}

\item{xreg}{external regressors used}
}
\value{
list of data.frames holding forecasting results
}
\description{
make test forecasts for further evaluation
}
