% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_model_identification.R
\name{make_train_test}
\alias{make_train_test}
\title{create training and test data for model selection}
\usage{
make_train_test(df, h.test, h)
}
\arguments{
\item{df}{data.frame holding all variables}

\item{h.test}{number of test periods}

\item{h}{forecast horizon}
}
\value{
nested list with training data for each test step
}
\description{
create training and test data for model selection
}
