% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Title}
\usage{
prepare_data(
  data = available_data,
  h = 6,
  h.test = 12,
  response = "Umsatz",
  reference = NULL,
  k = 24,
  no_lags = c("holidays", "workdays"),
  baseline_model = "workdays"
)
}
\arguments{
\item{data}{a \strong{named} list of data.frames with all data that should be included in the analysis}

\item{h.test}{how many periods should be considered for testing}

\item{response}{string corresponding to the name of the data.frame in data which contains the data for the response}

\item{reference}{optional. Reference against which the model should be evaluated. Default is 0.}

\item{k}{number of lags to consider for cross correlation analysis.}

\item{no_lags}{optional. Column names of variables where lags should be ignored.}

\item{baseline_model}{}
}
\value{
an object of class business.data
}
\description{
Title
}
