% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_forecast.R
\name{prepare_data_fc}
\alias{prepare_data_fc}
\title{Title}
\usage{
prepare_data_fc(
  response = models@data@response,
  model = models@extended[[1]]$model,
  data = object@data,
  step = 1
)
}
\arguments{
\item{data}{}
}
\description{
Title
}
