% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_model_identification.R
\name{reduce_mspace}
\alias{reduce_mspace}
\title{Title}
\usage{
reduce_mspace(
  traintest,
  xreg_candidates,
  response,
  exclude = exclude_var,
  spare_cores,
  baseline_m = baseline_model
)
}
\arguments{
\item{traintest}{data for training and testing models.}

\item{xreg_candidates}{names of variables that should be used as predictors}

\item{response}{response variable}

\item{baseline_m}{}
}
\description{
Title
}
